added auth

+++ main_app/views.py
@@ -1,9 +1,65 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
 from .models import Post, Topic
 from .forms import PostForm
+from django.http import HttpResponse
 
-# from django.http import HttpResponse
+
+def login_view(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST["username"].lower()
+        password = request.POST["password"]
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username or password is incorrect")
+
+    context = {"page": page}
+
+    return render(request, "main_app/login_and_register.html", context)
+
+
+def logout_view(request):
+    logout(request)
+    return redirect("login")
+
+
+def register_view(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(
+                request, "An error occurred during registration"
+            )
+
+    context = {"form": form}
+
+    return render(request, "main_app/login_and_register.html", context)
 
 
 def home(request):
@@ -39,6 +95,7 @@ def post(request, post_id):
     return render(request, "main_app/post.html", context)
 
 
+@login_required(login_url="login")
 def create_post(request):
     form = PostForm()
 
@@ -52,9 +109,13 @@ def create_post(request):
     return render(request, "main_app/post_form.html", context)
 
 
+@login_required(login_url="login")
 def update_post(request, post_id):
     post = Post.objects.get(id=post_id)
     form = PostForm(instance=post)
+
+    if request.user != post.owner:
+        return HttpResponse("You are not the owner of this post")
 
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
@@ -66,8 +127,12 @@ def update_post(request, post_id):
     return render(request, "main_app/post_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_post(request, post_id):
     post = Post.objects.get(id=post_id)
+
+    if request.user != post.owner:
+        return HttpResponse("You are not the owner of this post")
 
     if request.method == "POST":
         post.delete()
